+++ R/expectation.R
@@ -1,14 +1,35 @@
-#' Expectation class.
+#' The building block of all `expect_` functions
 #'
-#' You generate an expectation object with `expectation()`, and signal
-#' an expectation with `expect()`. If you're making your own expectation
-#' function, you'll typically call `expect(condition, "failure message")`.
-#' See the source code for the built-in expectations for details.
+#' Use this if you are writing your own expectation. See
+#' `vignette("custom-expectation")` for details
 #'
-#' @param passed a single logical value indicating whether the test passed
-#'  (`TRUE`), failed (`FALSE`), or threw an error (`NA`)
-#' @param message A text description of failure
-#' @param srcref Source reference, if known
+#' @param ok Was the expectation successful?
+#' @param failure_message What message should be shown if the expectation was
+#'   not successful?
+#' @param info Additional information. Included for backward compatibility
+#'   only and new expectations should not use it.
+#' @param srcref Only needed in very rare circumstances where you need to
+#'   forward a srcref captured elsewhere.
+#' @export
+expect <- function(ok, failure_message, info = NULL, srcref = NULL) {
+  exp <- as.expectation.logical(ok, failure_message, info = info, srcref = srcref)
+
+  withRestarts(
+    if (expectation_broken(exp)) {
+      stop(exp)
+    } else {
+      signalCondition(exp)
+    },
+    continue_test = function(e) NULL
+  )
+
+  invisible(exp)
+}
+
+#' Construct an expectation object
+#'
+#' For advanced use only.
+#'
 #' @keywords internal
 #' @export
 expectation <- function(type, message, srcref = NULL) {
@@ -29,29 +50,42 @@ expectation <- function(type, message, srcref = NULL) {
   )
 }
 
-#' @rdname expectation
-#' @export
-expect <- function(exp, ..., srcref = NULL) {
-  exp <- as.expectation(exp, ..., srcref = srcref)
-
-  withRestarts(
-    if (expectation_broken(exp)) {
-      stop(exp)
-    } else {
-      signalCondition(exp)
-    },
-    continue_test = function(e) NULL
-  )
-
-  invisible(exp)
-}
-
 #' @export
 #' @rdname expectation
-#' @param quo,label `make_label` is used to generate informative labels for
-#'   expressions. `quo` should be created with [rlang::enquo()], making it
-#'   possible to use unquoting to generate informative labels when running
-#'   multiple tests in a function or for loop.
+#' @param x object to test for class membership
+is.expectation <- function(x) inherits(x, "expectation")
+
+
+
+#' Quasi-labelling
+#'
+#' The first argument to every `expect_` function can use unquoting to
+#' construct better labels. This makes it easy to create informative labels
+#' expectations are used inside a function or a for loop. `quasi_label()` wraps
+#' up the details, returning the expression and label.
+#'
+#' @param quo A quosure created by `rlang::enquo()`.
+#' @param label An optional label to override the default. This is
+#'   only provided for internal usage. Modern expectations should not
+#'   include a `label` parameter.
+#' @keywords internal
+#' @return A list containing two elements:
+#' \item{val}{The evaluate value of `quo`}
+#' \item{lab}{The quasiquoted label generated from `quo`}
+#' @export
+#' @examples
+#' f <- function(i) if (i > 3) i * 9 else i * 10
+#' i <- 10
+#'
+#' # This short of expression commonly occurs inside a for loop or function
+#' # And the failure isn't helpful because you can't see the value of i
+#' # that caused the problem:
+#' show_failure(expect_equal(f(i), i * 10))
+#'
+#' # To overcome this issue, testthat allows you to unquote expressions using
+#' # !!. This causes the failure message to show the value rather than the
+#' # variable name
+#' show_failure(expect_equal(f(!!i), !!(i * 10)))
 quasi_label <- function(quo, label = NULL) {
   list(
     val = eval_bare(get_expr(quo), get_env(quo)),
@@ -152,7 +186,8 @@ as.expectation.expectation <- function(x, ..., srcref = NULL) {
 #' @export
 as.expectation.logical <- function(x, message, ..., srcref = NULL, info = NULL) {
   type <- if (x) "success" else "failure"
-  expectation(type, add_info(message, info), srcref = srcref)
+  message <- if (x) "success" else add_info(message, info)
+  expectation(type, message, srcref = srcref)
 }
 
 #' @export
@@ -185,11 +220,6 @@ as.expectation.skip <- function(x, ..., srcref = NULL) {
 
   expectation("skip", msg, srcref)
 }
-
-#' @export
-#' @rdname expectation
-#' @param x object to test for class membership
-is.expectation <- function(x) inherits(x, "expectation")
 
 #' @export
 print.expectation <- function(x, ...) cat(format(x), "\n")